requirements.txt
python-dotenv
streamlit
streamlit_authenticator

# feedparser
python-dateutil
# docopt

# TODO
# brew install mongodb-community
pymongo

websocket-client
# websocket
requests

embit
bitcoin
bech32
# brew install autoconf automake libtool pkg-config
bip32
bip39

# nostr_sdk
nostr_protocol
nostr_sdk

# nostr # for fuck sake this project is not up-to-date on pypi
# git+https://github.com/jeffthibault/python-nostr.git # for fuck sake this code is broken!!!
git+https://github.com/callebtc/python-nostr.git # a là callebtc
# git+https://github.com/vicariousdrama/python-nostr.git # a là vic
